integration-tests/pyproject.toml
[project]
name = "integration-tests"
version = "0.1.0"
description = "Integration tests for CLP."
readme = "README.md"
authors = [
    { name = "YScope Inc.", email = "dev@yscope.com" }
]
requires-python = ">=3.10"

[project.scripts]
integration-tests = "integration_tests:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.16.0",
    "ruff>=0.11.12",
    "pytest>=8.4.1",
    "pytest-env>=1.1.5",
    "PyYAML>=6.0",
    "types-PyYAML>=6.0.12.20240808",
]

[tool.mypy]
strict = true
ignore_missing_imports = true

# Additional output
pretty = true
show_error_code_links = true
show_error_context = true
show_error_end = true

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN401",  # Allow using `Any` type for function signatures
    "COM812",  # Redundant and conflicts with ruff format
    "D203",  # No blank line before docstrings (D211)
    "D205",  # Breaks if summary is larger than one line due to wrapping or if no summary exists
    "D212",  # Enforce docstring summary line on the next line after quotes (D213)
    "D400",  # First line of docstrings may not end in period
    "D401",  # Docstrings should be written in present tense (not imperative)
    "D415",  # First line of docstrings may not end in a period, question mark, or exclamation point
    "FBT",  # Allow bool positional parameters since other value positions are allowed
    "FIX002",  # Allow todo statements
    "PERF401",  # Allow for loops when creating lists
    "PERF403",  # Allow for loops when creating dicts
    "S311",  # Allow usage of `random` package
    "S603",  # Automatically trust inputs of subprocess execution
    "SIM102",  # Allow collapsible if statements for readability
    "SIM300",  # Skip Yoda-condition format fixes
    "TD002",  # Author unnecessary for todo statement
    "TD003",  # Issue link unnecessary for todo statement
    "UP015",  # Explicit open modes are helpful
]
isort.order-by-type = false

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "S101",  # Allow usage of pytest `assert`
    "TC003",  # Ignore performance overhead of imports only used for type checking
]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 100
